Addressing maintainers feedback
